+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.18.1
 
+* __DebugAdapter.print(notstring) will print the object's `describe` lineitem
 
 ## 0.18.0